update and finish mundo 1 python

+++ exercicios/035/triangulo.py
@@ -1,6 +1,6 @@
-r1 = int(input('Digite a primeira reta: '))
-r2 = int(input('Digite a segunda reta: '))
-r3 = int(input('Digite a terceira reta: '))
+r1 = float(input('Digite a primeira reta: '))
+r2 = float(input('Digite a segunda reta: '))
+r3 = float(input('Digite a terceira reta: '))
 
 if r1 + r2 > r3:
   print('É possível formar um triangulo com esses valores')
